backend/requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
cffi==1.11.5
cryptography==2.3.1
dataclasses==0.6
Django==2.1.3
django-cors-headers==2.4.0
djongo==1.2.30
idna==2.7
import-string==0.1.0
Pillow==5.3.0
pkg-resources==0.0.0
pycparser==2.19
PyJWT==1.6.4
pymongo==3.7.2
pytz==2018.7
reportlab==3.5.9
six==1.11.0
sqlparse==0.2.4
